feat(audio): replace SDL audio with cpal
Audio and graphics share no device, so the queue moves off the renderer
and onto `PixState` directly. That removes the audio half of the SDL
dependency ahead of the wgpu migration, and gives the web a working
audio path for the first time.

The public types were shaped by `SDL_AudioSpec` and are reshaped here:

- `AudioFormatNum` was a re-export of `sdl2::audio::AudioFormatNum`,
  making an SDL trait a bound on every user callback. It is now a sealed
  trait implemented for the eight sample types the backend converts.
- `AudioFormat` dropped its endianness variants. cpal describes samples
  in native endianness, so `S16LSB` and `S16MSB` had no meaning to it.
- `AudioSpec` and `AudioSpecDesired` name a `sample_rate` and a
  `buffer_size` in frames, replacing `freq`, `samples` and a `size` in
  bytes.
- `audio_queued_size` and `audio_size` count samples rather than bytes,
  and are named `audio_queued_samples` and `audio_buffer_size` to make
  the change visible at the call site.

Conversion between a callback's sample type and the device's runs
through `f32`. A direct type-to-type mapping needs a conversion bound
for every pair, which leaks into every signature that mentions a
callback.

Three behaviours the device thread needs, taken from tetanes: silence
until a full buffer has accumulated, so startup does not garble; a
decaying tail rather than a step to zero on underrun, which would click;
and a re-prime after total starvation.

The queue grows on demand behind a mutex up to the same ceiling SDL
enforced. A fixed ring sized to the device buffer cannot hold the bulk
pushes `audio_queue` makes, and reserving the ceiling up front would
cost 32MB for applications that never queue a sample. The device thread
takes the lock with `try_lock` and falls back to the decaying tail, so
it never blocks on the application.

Device selection asks for the device default rather than its maximum
sample rate. Taking the maximum opened capture at 384kHz.

A capture device that advertises a configuration can still refuse the
stream, which is what the ALSA default device does under PipeWire. The
default is opened once to check, and the enumerated devices are tried
when it fails, because a failed open cannot be retried after the
application's callback has moved into the stream.

Also drops the `mixer` SDL feature, which nothing imported.

+++ src/engine.rs
@@ -911,20 +911,20 @@ impl EngineBuilder {
     }
 
     /// Set audio sample rate in Hz (samples per second). Defaults to device fallback sample rate.
-    pub fn audio_sample_rate(&mut self, sample_rate: i32) -> &mut Self {
+    pub fn audio_sample_rate(&mut self, sample_rate: u32) -> &mut Self {
         self.settings.audio_sample_rate = Some(sample_rate);
         self
     }
 
     /// Set number of audio channels (1 for Mono, 2 for Stereo, etc). Defaults to device fallback
     /// number of channels.
-    pub fn audio_channels(&mut self, channels: u8) -> &mut Self {
+    pub fn audio_channels(&mut self, channels: u16) -> &mut Self {
         self.settings.audio_channels = Some(channels);
         self
     }
 
     /// Set audio buffer size in samples. Defaults to device fallback sample size.
-    pub fn audio_buffer_size(&mut self, buffer_size: u16) -> &mut Self {
+    pub fn audio_buffer_size(&mut self, buffer_size: u32) -> &mut Self {
         self.settings.audio_buffer_size = Some(buffer_size);
         self
     }
@@ -1112,8 +1112,7 @@ impl Engine {
                     self.state.set_delta_time(start_time, time_since_last);
                     self.state.increment_frame(time_since_last)?;
 
-                    // Measured before the pacing sleep below, so the number is the work the
-                    // frame did rather than the interval it was scheduled at.
+                    // Sampled before the pacing sleep below. See `crate::bench`.
                     if let Some(active) = bench.as_mut() {
                         if active.record(start_time.elapsed()) {
                             active.report(self.state.vsync_enabled());
